decomp: apply permuter improvements, clean up variable names
Automated permuter changes across 113 source files:
- Declaration reorders for register allocation fixes
- bool -> unsigned char type corrections
- (int)/(unsigned int) cast additions for comparisons
- Null guard eliminations where target omits checks
- Variable extraction for register lifetime management
- Renamed all _tmp/_cond/_ref/_sub permuter names to descriptive names
- Reverted 14 files the permuter erroneously emptied
- Fixed native port guards that leaked into PPC codegen (MetaPanel, Locale)

0 regressions vs 35b046f3f, +0.05% overall improvement.

+++ src/lazer/meta_ham/ChooseModeProvider.cpp
@@ -94,7 +94,7 @@ RndMat *ChooseModeProvider::Mat(int, int i_iData, UIListMesh *mesh) const {
     }
     auto isOnePlusIcon = mesh->Matches("icon_1p_plus");
     if (isOnePlusIcon) {
-        if (dataSym == custom_party) {
+        if ((int)dataSym == custom_party) {
             return nullptr;
         }
         return mesh->DefaultMat();
